pyproject.toml
[tool.poetry]
name = "langchain_sql"
version = "0.1.0"
description = ""
authors = ["Alice Kuo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.7"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.8"
sqlalchemy = "^2.0.36"
langchain-google-cloud-sql-pg = "^0.11.0"
langgraph = "^0.2.53"

[tool.poetry.scripts]
langchainsql = "langchain_sql.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
